[MAPS] Modified default precision from int to float, Also modifed warning Level from 3 to 4"
If you want to use the additional precision, you will need to extract the maps

If you want to extract the maps using int precision, you can use the parameter -f

+++ src/tools/map-extractor/loadlib/adt.h
@@ -388,7 +388,7 @@ class adt_MH2O
             if (h->offsData2a)
                 { return *((uint64*)((uint8*)this + 8 + h->offsData2a)); }
             else
-                { return 0xFFFFFFFFFFFFFFFFLL; }
+                { return (uint64)0xFFFFFFFFFFFFFFFFLL; }
         }
 
 };
